Hopefully last bunch of gcc-3.4 adjustments. Compiles cleanly and works now.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@445 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/include/video.h
@@ -40,8 +40,8 @@ void init_dualmon(void);
   #error "Currently USE_DUALMON can't be used together with VIDEO_CHECK_DIRTY"
 #endif
 
-#define CURSOR_START(c) ((int)((cshape*)&c)->start)
-#define CURSOR_END(c)   ((int)((cshape*)&c)->end)
+#define CURSOR_START(c) (((cshape*)&c)->start)
+#define CURSOR_END(c)   (((cshape*)&c)->end)
 #define NO_CURSOR 0x0100
 
 /*
